// src/model/enums/FormFieldType.ts
// 基础字段类型枚举
export enum FormFieldType {
	TEXT = "text",
	TEXTAREA = "textarea",
	PASSWORD = "password",
	NUMBER = "number",
	DATE = "date",
	DATETIME = "datetime",
	TIME = "time",
	CHECKBOX = "checkbox",
	TOGGLE = "toggle",
	RADIO = "radio",
	SELECT = "select",
	PROPERTY_VALUE_SUGGESTION = "property_value_suggestion",
	FILE_LIST = "file_list",
}